GUI: Replace "MT-32" tab with "AdLib" tab
This is the first step in many of cleaning up the launcher GUI for audio options. 

My proposed design is that the MT-32 and MIDI tabs be combined, while a new AdLib tab is created to merge the Mixed AdLib/MIDI and AdLib emulator settings from the Audio and MIDI tabs. This will make it easier to make future changes, such as creating a universal "MIDI (GM/MT-32)" device under Audio device pop-up to replace the existing list of all MIDI devices that is combined with the list of emulators. It also means that if a change is made to disable GM devices when MT-32 devices are selected, the disabling effect will be more apparent since both devices will be on the same tab.